Numbers of sudoku board
Given a list of numbers in a file "numbers.txt", this program generate
a 9x9 game board similar to sudoku, however with only random numbers.

The program will find all the numbers in order from the file that are
on the board and display.

The program will then remove each of those numbers it finds on the game
board from the game board. This  process will stop after it can't find
any more numbers.