Binary Tree Constructor
Given a file on the command line (Ex. with "bitree" being an executable
file and "the dat.txt" the file, this will run via "./bitree the
dat.txt") this program will construct a binary tree.

The items in the file will have the format "something 10.2"
"something" - > species
"10.2" -> the species' score

The binary tree will be constructed based on the relation of the
species score. A diagram is attached to show this construction. The
result of the tree will be the the root of the tree in balance
parentheses form and the relational score.